refactor arrow loop animations to each component
