+++ components/BeingAtTorchbox/PageNav/PageNav.module.scss
@@ -18,11 +18,6 @@
     transform: rotate(10deg);
 }
 
-.svgArrowLoopPath {
-    stroke-dasharray: 1;
-    stroke-dashoffset: 1;
-}
-
 .navigation {
     display: flex;
     flex-direction: row;
