+++ components/SVG/ArrowLoop/ArrowLoopLeft.tsx
@@ -1,3 +1,7 @@
+import { animateWithOptions, drawSVGPath } from 'lib/animations';
+import { RefObject, useEffect, useRef } from 'react';
+import styles from './ArrowLoop.module.scss';
+
 type ArrowLoopLeftProps = {
     size?: number;
     className?: string;
@@ -6,20 +10,90 @@ type ArrowLoopLeftProps = {
 export const ArrowLoopLeft = ({
     size = 217,
     className,
-}: ArrowLoopLeftProps) => (
-    <svg
-        className={className}
-        width={size}
-        height={size}
-        viewBox="0 0 171 217"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-    >
-        <path
-            d="M169.491 38.0196C170.245 37.2124 170.202 35.9469 169.395 35.1929C168.587 34.4389 167.322 34.482 166.568 35.2892L169.491 38.0196ZM104.676 85.8013L103.799 84.0035L104.676 85.8013ZM51.9132 84.5255L52.5062 82.6154L51.9132 84.5255ZM2.95385 20.2958L0.967694 20.0609L2.95385 20.2958ZM10.6416 2.68937L9.88278 0.838901L10.6416 2.68937ZM22.1569 7.36055L21.0845 9.04871L22.1569 7.36055ZM43.1362 24.1775L41.6679 25.5354L43.1362 24.1775ZM67.5054 64.4608L65.6721 65.2601L67.5054 64.4608ZM82.4567 135.961L80.4602 135.842L82.4567 135.961ZM51.6672 214.808C51.704 215.912 52.6288 216.777 53.7327 216.74L71.7227 216.141C72.8267 216.104 73.6918 215.179 73.655 214.075C73.6182 212.971 72.6934 212.106 71.5895 212.143L55.5984 212.676L55.0654 196.685C55.0286 195.581 54.1038 194.716 52.9999 194.753C51.8959 194.789 51.0308 195.714 51.0676 196.818L51.6672 214.808ZM166.568 35.2892C148.952 54.1482 126.954 72.7175 103.799 84.0035L105.552 87.5991C129.298 76.0245 151.68 57.0871 169.491 38.0196L166.568 35.2892ZM103.799 84.0035C96.3248 87.6468 87.5207 88.5666 78.5051 87.9194C69.4999 87.2729 60.4229 85.0734 52.5062 82.6154L51.3202 86.4355C59.391 88.9414 68.7914 91.2324 78.2187 91.9091C87.6354 92.5851 97.2189 91.6609 105.552 87.5991L103.799 84.0035ZM52.5062 82.6154C24.1732 73.8184 1.3202 51.1424 4.94001 20.5306L0.967694 20.0609C-2.9676 53.3405 21.9811 77.3262 51.3202 86.4355L52.5062 82.6154ZM4.94001 20.5306C5.22465 18.1235 5.70546 14.4791 6.77758 11.177C7.89143 7.74644 9.43497 5.34573 11.4003 4.53985L9.88278 0.838901C6.15135 2.36894 4.13627 6.35925 2.97308 9.94179C1.76817 13.6529 1.25352 17.6438 0.967694 20.0609L4.94001 20.5306ZM11.4003 4.53985C11.6563 4.4349 12.1689 4.37931 13.0565 4.60528C13.9146 4.82376 14.9023 5.2492 15.9405 5.80109C18.0711 6.93371 19.9019 8.2975 21.0845 9.04871L23.2293 5.67238C22.3889 5.13847 20.1136 3.48945 17.8181 2.26914C16.6432 1.64457 15.3379 1.05852 14.0433 0.728943C12.7782 0.406844 11.2785 0.266605 9.88278 0.838901L11.4003 4.53985ZM21.0845 9.04871C28.4799 13.7467 35.8019 19.1922 41.6679 25.5354L44.6046 22.8196C38.41 16.121 30.7693 10.4622 23.2293 5.67238L21.0845 9.04871ZM41.6679 25.5354C52.1543 36.8749 59.4437 50.9742 65.6721 65.2601L69.3388 63.6615C63.0612 49.2628 55.5569 34.6628 44.6046 22.8196L41.6679 25.5354ZM65.6721 65.2601C75.6084 88.0506 81.9436 110.936 80.4602 135.842L84.4532 136.08C85.9877 110.315 79.4154 86.7739 69.3388 63.6615L65.6721 65.2601ZM80.4602 135.842C78.8475 162.921 70.8517 193.443 52.2056 213.375L55.1266 216.108C74.6916 195.194 82.8158 163.573 84.4532 136.08L80.4602 135.842Z"
-            fill="#251657"
-        />
-    </svg>
-);
+}: ArrowLoopLeftProps) => {
+    const containerRef: RefObject<SVGSVGElement> = useRef<SVGSVGElement | null>(
+        null,
+    );
+
+    // Animate the entrance of the SVG arrow loop.
+    // Prefers reduced motion styles are applied in `PageNav.module.scss` to prevent this animation.
+    useEffect(() => {
+        const containerNode = containerRef?.current;
+        const hasIOSupport = !!window.IntersectionObserver;
+        if (!hasIOSupport || !containerNode) return;
+
+        const callback = (
+            entries: IntersectionObserverEntry[],
+            observer: IntersectionObserver,
+        ) => {
+            // Look at all intersection observer event entries reported
+            entries.forEach((entry) => {
+                // If the event entry is of a target element intersecting with the observer
+                if (entry.isIntersecting) {
+                    const pathElements = entry.target.children;
+
+                    // Emotion: airy, graceful. Avoid the second path feeling like a tick box check. Draw the user down the page
+                    animateWithOptions(
+                        pathElements[0] as SVGElement,
+                        drawSVGPath,
+                        {
+                            duration: 1700,
+                            easing: 'cubic-bezier(0.45, 0, 0.55, 1)',
+                        },
+                    );
+                    animateWithOptions(
+                        pathElements[1] as SVGElement,
+                        drawSVGPath,
+                        { delay: 1700, duration: 500, easing: 'ease' },
+                    );
+
+                    // Stop observing so the animation doesn't replay if they scroll down again.
+                    observer.unobserve(entry.target);
+                }
+            });
+        };
+
+        let observerThreshold = 0.1;
+
+        const observer = new IntersectionObserver(callback, {
+            root: null,
+            rootMargin: '0px',
+            threshold: observerThreshold,
+        });
+
+        observer.observe(containerNode);
+
+        return () => observer.disconnect();
+    });
+
+    return (
+        <svg
+            className={className}
+            width={size}
+            height={size}
+            viewBox="0 0 171 217"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth="3.5"
+            stroke="#251657"
+            ref={containerRef}
+        >
+            <path
+                className={styles.svgArrowLoopPath}
+                pathLength="1"
+                stroke="#251657"
+                d="M169.491 38.0196C95 140-28 71 5.484 9.235 9.693 1.621 12.69 1.776 17.621 2.872 51 8 123 124 55 214"
+            />
+            <path
+                className={styles.svgArrowLoopPath}
+                pathLength="1"
+                stroke="#251657"
+                d="M52.8 197.184 54.084 214.904 70.378 213.0"
+            />
+        </svg>
+    );
+};
 
 export default ArrowLoopLeft;
